backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.35
asyncpg==0.30.0
aiosqlite==0.20.0
redis==5.0.8
celery==5.4.0
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
python-multipart==0.0.9
aiohttp==3.10.5
beautifulsoup4==4.12.3
lxml==5.3.0
pytesseract==0.3.13
Pillow==10.4.0
PyPDF2==3.0.1
pdfplumber==0.11.0
python-docx==1.1.2
openpyxl==3.1.5
pandas==2.2.2
numpy==1.26.4
reportlab==4.2.2
tiktoken==0.7.0
sentence-transformers==3.0.1
openai==1.51.0
prometheus-client==0.20.0
loguru==0.7.2
email-validator==2.1.0
pytz==2024.1